Something wrong with initializing with double.

+++ IceRink.h
@@ -2,7 +2,7 @@
 
 class IceRink{
 	public:
-		double x_size, y_size;
+		int x_size, y_size;
 		IceRink(double x_size, double y_size);
 		struct Gate{
 			double x_coord, y_coord, width, height;
